added prop types
