+++ config/webpack.config.dev.js
@@ -101,23 +101,6 @@ module.exports = {
   module: {
     strictExportPresence: true,
     rules: [
-      {
-        test: /\.svg$/,
-        use: [
-          {
-            loader: 'babel-loader',
-            query: {
-              presets: ['airbnb'],
-            },
-          },
-          {
-            loader: 'react-svg-loader',
-            query: {
-              jsx: true,
-            },
-          },
-        ],
-      },
       // TODO: Disable require.ensure as it's not a standard language feature.
       // We are waiting for https://github.com/facebookincubator/create-react-app/issues/2176.
       // { parser: { requireEnsure: false } },
