+++ src/client/components/region-banner-component/index.jsx
@@ -1,73 +1,10 @@
 import React, { Component } from 'react';
-import { Card, Icon, Image, Grid, Container } from 'semantic-ui-react';
+import { Card, Grid, Container } from 'semantic-ui-react';
+import PropTypes from 'prop-types';
+
 import 'client/components/region-banner-component/index.scss';
-import birka from 'client/components/region-banner-component/img/birka.svg'
 
-/* {
-    id: number,
-    city: string,
-    flagUrl: string,
-    properties: number,
-    avgPrice: number
-    pictureUrl: string
-} */
-
-class BannerComponent extends React.Component {
-    constructor(props) {
-        super(props);
-        console.log('ctor banner component')
-    }
-
-    render() {
-        const { cityInfo, onCardClick } = this.props;
-        const imgStyles = (url) => ({
-            background: `url(${url})`,
-            backgroundSize: 'cover',
-
-        });
-
-        console.log('This is birka >>>> ', birka);
-
-        return (
-            // <Container onClick={onCardClick} className="banner">
-            //     <div className="banner-header">
-            //         <h4 className="title  flag-icon flag-icon-ua">{cityInfo.city}
-            //             <img src={cityInfo.flagUrl} alt="/" style={{ width: 59 }} />
-            //         </h4>
-
-            //         <h5 className="subtitle">{cityInfo.properties}   properties</h5>
-            //     </div>
-            //     <div className="banner-extra">
-            //         <div className="avg-price">Average price UAH{cityInfo.avgPrice}</div>
-            //     </div>
-            // </Container>
-
-            <Container onClick={onCardClick} className="banner" style={imgStyles(cityInfo.pictureUrl)}>
-                <Card.Content>
-                    <Card.Header className="title">
-                        {cityInfo.city}
-                        <img src={cityInfo.flagUrl} alt="/" style={{ width: 59 }} />
-                    </Card.Header>
-                    <Card.Meta>
-                        <h5 className="subtitle">{Intl.NumberFormat('en-US').format(cityInfo.properties)}   properties</h5>
-                    </Card.Meta>
-                    <Card.Description className="avg-price" style={{ backgroundImage: `url(${birka})` }}>
-                        <span className="avg-price_text">
-                            Average price
-                        </span>
-                        <span className="avg-price_count">
-                            UAH  {Intl.NumberFormat('en-US').format(cityInfo.avgPrice)}
-                        </span>
-
-                    </Card.Description>
-                </Card.Content>
-
-            </Container >
-
-        );
-    }
-}
-
+import { BannerComponent } from './banner';
 
 export default class BannerListComponent extends React.Component {
     constructor(props) {
@@ -75,8 +12,8 @@ export default class BannerListComponent extends React.Component {
     }
 
     onCardClick = (cityId) => {
-        // awsome logic ...
-        console.log(cityId);
+        // some logic ...
+
     }
 
     render() {
@@ -86,28 +23,7 @@ export default class BannerListComponent extends React.Component {
             city3, city4, city5
         ] = cityInfos;
 
-
-
         return (
-            // <Grid celled>
-            //     <Grid.Column width={8} style={imgStyles(city1.pictureUrl)}>
-            //         <BannerComponent cityInfo={city1} onClick={() => this.onCardClick(city1.id)} />
-            //     </Grid.Column>
-            //     <Grid.Column width={8} style={imgStyles(city2.pictureUrl)}>
-            //         <BannerComponent cityInfo={city2} onClick={() => this.onCardClick(city2.id)} />
-            //     </Grid.Column>
-
-            //     < Grid.Column width={5} style={imgStyles(city3.pictureUrl)}>
-            //         <BannerComponent cityInfo={city3} onClick={() => this.onCardClick(city3.id)} />
-            //     </Grid.Column>
-            //     <Grid.Column width={6} style={imgStyles(city4.pictureUrl)}>
-            //         <BannerComponent cityInfo={city4} onClick={() => this.onCardClick(city4.id)} />
-            //     </Grid.Column>
-            //     <Grid.Column width={5} style={imgStyles(city5.pictureUrl)}>
-            //         <BannerComponent cityInfo={city5} onClick={() => this.onCardClick(city5.id)} />
-            //     </Grid.Column>
-            // </Grid>
-
             <Grid >
                 <Grid.Row columns={2}>
                     <Grid.Column>
@@ -134,3 +50,16 @@ export default class BannerListComponent extends React.Component {
 
     }
 }
+
+BannerListComponent.propTypes = {
+    cityInfos: PropTypes.arrayOf(
+        PropTypes.shape({
+            id: PropTypes.number.isRequired,
+            city: PropTypes.string.isRequired,
+            flagUrl: PropTypes.string.isRequired,
+            properties: PropTypes.number.isRequired,
+            avgPrice: PropTypes.number.isRequired,
+            pictureUrl: PropTypes.string.isRequired
+        })
+    ).isRequired
+};
